Alteração de campos de valor de double para decimal, inclusão de validator para ReturnCar

+++ CarRent.API/Application/Commands/Requests/CarCommands/CreateCarCommand.cs
@@ -3,5 +3,5 @@ using MediatR;
 
 namespace CarRent.API.Application.Commands.Requests.CarCommands
 {
-    public record CreateCarCommand(string Model, string Maker, int Year, double DailyPrice) : IRequest<Car> { }
+    public record CreateCarCommand(string Model, string Maker, int Year, decimal DailyPrice) : IRequest<Car> { }
 }
